Separated draggablePositionItem to independent component
